% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byyr.R
\name{abundance_byyr}
\alias{abundance_byyr}
\alias{basal_area_byyr}
\title{Create tables of abundance and basal area by year.}
\usage{
abundance_byyr(vft, ...)

basal_area_byyr(vft, ...)
}
\arguments{
\item{vft}{A ForestGEO-like dataframe; particularly a ViewFullTable. As such,
it should contain columns `PlotName`, `CensusID`, `TreeID`, `StemID`,
`Status`, `DBH`, `Genus`, `SpeciesName`, `ExactDate`, `PlotCensusNumber`,
`Family`, `Tag`, and `HOM`. `ExactDate` should contain dates from
1980-01-01 to the present day in the format yyyy-mm-dd.}

\item{...}{Expressions to pick main stems of a specific `dbh` range.}
}
\value{
A dataframe.
}
\description{
* `abundance_byyr()` first picks the main stem of each tree (see
?[fgeo.tool::pick_main_stem()] and then, for each species and each
(round mean) year of measurement, counts the number of
trees. The result includes __main stems__ within a given dbh range.
* `basal_area_byyr()` first sums the basal basal area of all stems of each
tree, and then, for each species and each (round mean) year of measurement,
sums the basal area of all trees. The result includes all stems within a
given dbh range (notice the difference with `abundance_byyr()`).
}
\details{
You don't need to pick stems by status before feeding data to these
functions. Doing so may make your code more readable but it should not affect
the result. This is because the expressions passed to `...` pick data by
`dbh` and exclude missing the `dbh` values associated to non-alive stems,
including dead, missing, and gone.
}
\examples{
library(fgeo.tool)

vft <- example_byyr
vft
abundance_byyr(vft, DBH >= 10, DBH < 20)
abundance_byyr(vft, DBH >= 10)
basal <- basal_area_byyr(vft, DBH >= 10)
basal

\dontrun{
# Convert units
# This exmaple requires the following packages (see `?install.packages`):
required_packages <- c("measurements", "purrr")
not_installed <- !all(sapply(required_packages, requireNamespace, quietly = T))
if (not_installed) {
  stop("To run this section please install the required packages.")
} else {
  years <- c("yr_2001", "yr_2002")
  basal_he <- purrr::modify_at(
    basal,
    .at = years,
    .f =  ~measurements::conv_unit(.x, from = "mm2", to = "hectare")
  )
  basal_he

  # Standardize
  number_of_hectares <- 50
  purrr::map_at(
    basal_he,
    .at = years,
    .f = ~.x / number_of_hectares
  )
}
}
}
\seealso{
[fgeo.tool::pick_main_stem()], [fgeo.tool::convert_unit()].

Other functions for abundance and basal area: \code{\link{abundance}}
}
\concept{functions for abundance and basal area}
