% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceperquad2.R
\name{abundanceperquad2}
\alias{abundanceperquad2}
\title{A version of ctfs::abundanceperquad() focused only for abundance.}
\usage{
abundanceperquad2(censdata, mindbh = 10, plotdim = c(1000, 500),
  gridsize = 100, type = "abund", dbhunit = "mm")
}
\arguments{
\item{censdata}{A table of plot census data.}

\item{mindbh}{The minimum diameter above which the counts are done. Trees
smaller than `mindbh` are excluded. If `NULL`, all living trees are
included.}

\item{plotdim}{The x and y dimensions of the plot.}

\item{gridsize}{Side of the square quadrate.}

\item{type}{"abund". Other options are intentionally not possible.}

\item{dbhunit}{'cm' or 'mm', only used for basal area}
}
\description{
A version of ctfs::abundanceperquad() focused only for abundance.
}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{calcslope}},
  \code{\link{findborderquads}}, \code{\link{quadslope}}
}
\author{
Rick Condit, Suzanne Lao.
}
\concept{functions from http://ctfs.si.edu/Public/CTFSRPackage/}
\keyword{internal}
