% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as_tibble.demography_ctfs}
\alias{as_tibble.demography_ctfs}
\alias{as.data.frame.demography_ctfs}
\title{Coerce objects of class "demography_ctfs" to "data.frame" and "tbl" (tibble).}
\usage{
\method{as_tibble}{demography_ctfs}(x, ...)

\method{as.data.frame}{demography_ctfs}(x, ...)
}
\arguments{
\item{x}{An object of class demography_ctfs.}

\item{...}{Not used.}
}
\value{
A (tibble) dataframe.
}
\description{
Coerce objects of class "demography_ctfs" to "data.frame" and "tbl" (tibble).
}
\examples{
census1 <- fgeo.x::tree5
census2 <- fgeo.x::tree6
by_sp_and_quadrat <- interaction(census1$sp, census1$quadrat)

demography_result <- recruitment_ctfs(
  census1,
  census2,
  split1 = by_sp_and_quadrat
)

as_tibble(demography_result)
}
\seealso{
\code{\link[=as_tibble]{as_tibble()}}.

Other demography functions: \code{\link{recruitment_ctfs}}

Other methods for common generics: \code{\link{as_tibble.tt_lst}},
  \code{\link{summary.tt_df}}
}
\concept{demography functions}
\concept{methods for common generics}
