% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as_tibble.tt_lst}
\alias{as_tibble.tt_lst}
\alias{as.data.frame.tt_lst}
\title{Coerce objects of class "tt_lst" to "data.frame" and "tbl" (tibble).}
\usage{
\method{as_tibble}{tt_lst}(x, ...)

\method{as.data.frame}{tt_lst}(x, ...)
}
\arguments{
\item{x}{An object of class tt_lst.}

\item{...}{Not used.}
}
\value{
A dataframe.
}
\description{
This method creates a dataframe from the output of \code{tt_test()}
(which is a list of class "tt_lst").
}
\examples{
census <- fgeo.x::tree6_3species
habitat <- fgeo.x::habitat

as_tibble(
  tt_test(census, habitat)
)
}
\seealso{
\code{\link[=as_tibble]{as_tibble()}}.

Other habitat functions: \code{\link{fgeo_habitat}},
  \code{\link{fgeo_topography}}, \code{\link{tt_test}}

Other methods for common generics: \code{\link{as_tibble.demography_ctfs}},
  \code{\link{summary.tt_df}}
}
\concept{habitat functions}
\concept{methods for common generics}
