% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_topography.R
\name{calcslope}
\alias{calcslope}
\title{Internal.}
\usage{
calcslope(z, gridsize)
}
\description{
Internal.
}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{abundanceperquad2}},
  \code{\link{findborderquads}}, \code{\link{quadslope}}
}
\author{
Rick Condit, Suzanne Lao.
}
\concept{functions from http://ctfs.si.edu/Public/CTFSRPackage/}
\keyword{internal}
