% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_crucial_names.R
\name{check_crucial_names}
\alias{check_crucial_names}
\title{Check if an object contains specific names.}
\usage{
check_crucial_names(x, nms)
}
\arguments{
\item{x}{A named object.}

\item{nms}{String; names expected to be found in `x`.}
}
\value{
Invisible `x`, or an error with informative message.
}
\description{
Check if an object contains specific names.
}
\examples{
v <- c(x = 1)
check_crucial_names(v, "x")

dfm <- data.frame(x = 1)
check_crucial_names(dfm, "x")
}
\concept{exported functions for internal use}
\concept{functions for developers}
\concept{functions to check inputs.}
\concept{general functions to assert}
\keyword{internal}
