% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_byyr.R
\docType{data}
\name{example_byyr}
\alias{example_byyr}
\title{Example dataset with some columns of a ViewFullTable.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 13 columns.}
\usage{
example_byyr
}
\description{
Example dataset with some columns of a ViewFullTable.
}
\examples{
library(tibble)

example_byyr
}
\keyword{datasets}
