% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.analyze-package.R
\docType{package}
\name{fgeo.analyze-package}
\alias{fgeo.analyze}
\alias{fgeo.analyze-package}
\title{fgeo.analyze: Analyze ForestGEO Data}
\description{
Analyze ForestGEO data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/fgeo.analyze}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo.analyze/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Authors:
\itemize{
  \item Gabriel Arellano \email{gabriel.arellano.torres@gmail.com} [reviewer]
  \item  Richard  Condit \email{richardcondit@gmail.com}
  \item Matteo Detto \email{teoparaglider@gmail.com}
  \item Kyle Harms \email{ kharms@lsu.edu}
  \item Suzanne Lao \email{laoz@si.edu} [reviewer]
  \item Sabrina Russo \email{srusso2@unl.edu} [reviewer]
  \item Daniel Zuleta \email{dfzuleta@gmail.com} [reviewer]
}

Other contributors:
\itemize{
  \item KangMin Ngo \email{ngokangmin@gmail.com} [reviewer]
  \item Haley Overstreet \email{OverstreetH@si.edu} [reviewer]
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder]
}

}
\keyword{internal}
