% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_habitat.R
\name{fgeo_habitat}
\alias{fgeo_habitat}
\title{Structure habitat data from measures of topography.}
\usage{
fgeo_habitat(elevation, gridsize, n, ...)
}
\arguments{
\item{elevation}{One of these:
* A list with at least three elements: `col` containing
elevation data; and `xdim` and `ydim` giving plot dimensions; OR
* A dataframe containing elevation data, in which
case the parameters `xdim` and `ydim` must be provided. It may be the
element `col` of a ForestGEO elevation-list or an object of class
fgeo_elevation (see [fgeo.tool::fgeo_elevation()]).}

\item{gridsize}{Number giving the size of each quadrat for which a habitat
is calculated. Commonly, `gridsize = 20`.}

\item{n}{Integer. Number of cluster-groups to construct (passed to the
argument `k` to [stats::cutree()]).}

\item{...}{Arguments passed to [fgeo_topography()].}
}
\value{
A dataframe of subclass fgeo_habitat, with columns `gx` and `gy`,
  rounded with accuracy determined by `gridsize`, and column `habitats`, with
  as many distinct integer values as determined by the argument `n`.
}
\description{
This function constructs habitat data based on elevation data. It calculates
habitats in two steps:
1. It calculates mean elevation, convexity and slope for each quadrat (via
[fgeo_topography()])).
2. It calculates habitats based on hierarchical clustering of the topographic
metrics from step 1 (via [stats::hclust()]).
}
\section{Input}{


The main input can be either the elevation list that ForestGEO delivers, or
the element \code{col} of such list -- which is a dataframe containing the
elevation data. Notice that the required arguments vary according to the main
input (the elevation list or the elevation dataframe). Whatever the input,
the dataframe containing the elevation data must have columns \code{gx} and \code{gy}
or \code{x} and \code{y}.

}

\examples{
# Input a ForestGEO-like elevation list
elev_list <- fgeo.x::elevation
habs <- fgeo_habitat(elev_list, n = 4, gridsize = 20)
str(habs)

\dontrun{
if (requireNamespace("fgeo.map", quietly = TRUE)) {
  library(fgeo.map)
  autoplot(habs)
}
}
# A good use of habitat data is for calculating species-habitat associations
elev_list <- fgeo.x::elevation
habitat <- fgeo_habitat(elev_list, gridsize = 20, n = 4)
census <- fgeo.x::tree6_3species
species <- unique(census$sp)
to_df(
  tt_test(census, species, habitat)
)

# If elevation is not a list but a dataframe, you must provide xdim and ydim
elev_df <- fgeo.x::elevation$col
hab2 <- fgeo_habitat(elev_df, gridsize = 20, n = 4, xdim = 320, ydim = 500)
str(hab2)
}
\seealso{
[fgeo.map::autoplot.fgeo_habitat()], [fgeo_topography()].

Other habitat functions: \code{\link{fgeo_topography}},
  \code{\link{summary.tt_lst}}, \code{\link{to_df.tt_lst}},
  \code{\link{tt_test}}

Other functions to construct fgeo classes: \code{\link{fgeo_topography}}
}
\author{
Rick Condit.
}
\concept{functions to construct fgeo classes}
\concept{habitat functions}
