% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_topography.R
\name{fgeo_topography}
\alias{fgeo_topography}
\alias{fgeo_topography.data.frame}
\alias{fgeo_topography.list}
\title{Structure topography data with mean elevation, convexity, and slope.}
\usage{
fgeo_topography(elevation, ...)

\method{fgeo_topography}{data.frame}(elevation, gridsize, xdim = NULL,
  ydim = NULL, edgecorrect = TRUE, ...)

\method{fgeo_topography}{list}(elevation, gridsize, edgecorrect = TRUE,
  ...)
}
\arguments{
\item{elevation}{One of these:
\itemize{
\item A list with at least three elements: \code{col} containing
elevation data; and \code{xdim} and \code{ydim} giving plot dimensions; OR
\item A dataframe containing elevation data, in which
case the parameters \code{xdim} and \code{ydim} must be provided. It may be the
element \code{col} of a ForestGEO elevation-list or an object of class
fgeo_elevation (see \code{\link[fgeo.tool:fgeo_elevation]{fgeo.tool::fgeo_elevation()}}).
}}

\item{...}{Other arguments passed to methods.}

\item{gridsize}{Number giving the size of each quadrat for which a habitat
is calculated. Commonly, \code{gridsize = 20}.}

\item{xdim, ydim}{(If \code{elevation} is a dataframe) \code{x} and \code{y} dimensions of
the plot.}

\item{edgecorrect}{Correct convexity in edge quadrats?}
}
\value{
A dataframe of subclass fgeo_topography.
}
\description{
Structure topography data with mean elevation, convexity, and slope.
}
\section{Acknowledgment}{

Thanks to Jian Zhang for reporting a bug (issue 59).
}

\examples{
elev_list <- fgeo.x::elevation
fgeo_topography(elev_list, gridsize = 20)

elev_df <- elev_list$col
fgeo_topography(elev_df, gridsize = 20, xdim = 320, ydim = 500)

}
\seealso{
\code{\link[=fgeo_habitat]{fgeo_habitat()}}.

Other habitat functions: \code{\link{fgeo_habitat}},
  \code{\link{tt_test}}

Other functions to construct fgeo classes: \code{\link{fgeo_habitat}}
}
\author{
This function wraps code by Rick Condit.
}
\concept{functions to construct fgeo classes}
\concept{habitat functions}
