% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{luquillo_top3_sp}
\alias{luquillo_top3_sp}
\title{A tiny dataset from Luquillo of the three most abundant species.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 434 rows and 19 columns.}
\usage{
luquillo_top3_sp
}
\description{
A tiny dataset from Luquillo of the three most abundant species.
}
\examples{
str(luquillo_top3_sp)
}
\keyword{datasets}
