% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_topography.R
\name{quadslope}
\alias{quadslope}
\title{Internal.}
\usage{
quadslope(cornerelev, gridsize)
}
\description{
Internal.
}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{abundanceperquad2}},
  \code{\link{calcslope}}, \code{\link{findborderquads}}
}
\author{
Rick Condit, Suzanne Lao.
}
\concept{functions from http://ctfs.si.edu/Public/CTFSRPackage/}
\keyword{internal}
