% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_tibble}
\title{Coerce objects of different fgeo classes to "data.frame" and "tbl" (tibble).}
\arguments{
\item{x}{An fgeo object of supported class.}

\item{...}{Other arguments passed to methods.}
}
\value{
\preformatted{* `as.data.drame` returns an object of class "data.frame"
* `as_tibble` returns an object of class "tibble".
}
}
\seealso{
\code{\link[base:as.data.frame]{base::as.data.frame()}}, \code{\link[=as_tibble]{as_tibble()}}.
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

