% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-tt_lst.R
\name{summary.tt_lst}
\alias{summary.tt_lst}
\title{Summary objects of class "tt_lst".}
\usage{
\method{summary}{tt_lst}(object, ...)
}
\arguments{
\item{object}{The output of \code{\link[=tt_test]{tt_test()}}, i.e. an S3 object of class tt_lst.}

\item{...}{Other arguments passed to methods.}
}
\value{
A dataframe.
}
\description{
This method helps you to interpret the results of \code{tt_test()} (which
outputs objects of class "tt_lst").
}
\examples{
census <- fgeo.x::tree6_3species
habitat <- fgeo.x::habitat

summary(
  tt_test(census, habitat)
)

}
\seealso{
\code{\link[=tt_test]{tt_test()}}.

Other habitat functions: \code{\link{fgeo_habitat}},
  \code{\link{fgeo_topography}},
  \code{\link{to_df.tt_lst}}, \code{\link{tt_test}}
}
\author{
Daniel Zuleta.
}
\concept{habitat functions}
\concept{methods for common generics}
