% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_df.R
\name{to_df.demography_ctfs}
\alias{to_df.demography_ctfs}
\title{Dataframe objects of class "demography_ctfs".}
\usage{
\method{to_df}{demography_ctfs}(.x, ...)
}
\arguments{
\item{.x}{An object of class demography_ctfs.}

\item{...}{Other arguments passed to \code{to_df()}.}
}
\value{
A (tibble) dataframe.
}
\description{
Dataframe objects of class "demography_ctfs".
}
\examples{
census1 <- fgeo.x::tree5
census2 <- fgeo.x::tree6
by_sp_and_quadrat <- interaction(census1$sp, census1$quadrat)

demography_result <- recruitment_ctfs(
  census1,
  census2,
  split1 = by_sp_and_quadrat
)

to_df(demography_result)

}
\seealso{
\code{\link[=to_df]{to_df()}}.

Other demography functions: \code{\link{recruitment_ctfs}}

Other methods for fgeo generics: \code{\link{to_df.tt_lst}}
}
\concept{demography functions}
\concept{methods for fgeo generics}
