% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_df.R
\name{to_df}
\alias{to_df}
\title{Create objects of class "data.frame" from other fgeo classes.}
\usage{
to_df(.x, ...)
}
\arguments{
\item{.x}{An fgeo object of supported class.}

\item{...}{Other arguments passed to methods.}
}
\value{
A dataframe.
}
\description{
Most of the popular, general-purpose tools for data science input objects of
class "data.frame" (<https://www.tidyverse.org/>). However, several __fgeo__
functions (either inherited from the original CTFS R Package or contributed
by ForestGEO partners) output data of different class. Taking as input
different classes of __fgeo__ objects, `to_df()` provides a simple,
consistent way to create dataframes.
}
\details{
This generic provides methods for classes that cannot be correctly coerced
simply with [base::as.data.frame()] (or similar functions from the
__tibble__ package).
}
\seealso{
[to_df.krig_lst()].
}
\concept{fgeo generics}
