% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_df.R
\name{to_df.tt_lst}
\alias{to_df.tt_lst}
\title{Dataframe objects of class "tt_lst".}
\usage{
\method{to_df}{tt_lst}(.x, ...)
}
\arguments{
\item{.x}{An object of class tt_lst.}

\item{...}{Other arguments passed to [to_df()].}
}
\value{
A dataframe.
}
\description{
This method creates a dataframe from the output of `tt_test()`
(which is a list of class "tt_lst").
}
\examples{
cns <- fgeo.x::tree6_3species
hab <- fgeo.x::habitat
tt_lst <- tt_test(cns, unique(cns$sp), hab)
to_df(tt_lst)
}
\seealso{
[to_df()].

Other habitat functions: \code{\link{fgeo_habitat}},
  \code{\link{fgeo_topography}},
  \code{\link{summary.tt_lst}}, \code{\link{tt_test}}

Other methods for fgeo generics: \code{\link{to_df.demography_ctfs}}
}
\concept{habitat functions}
\concept{methods for fgeo generics}
