% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.analyze-package.R
\docType{package}
\name{fgeo.analyze-package}
\alias{fgeo.analyze}
\alias{fgeo.analyze-package}
\title{fgeo.analyze: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/fgeo.analyze}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo.analyze/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder]
}

}
\keyword{internal}
