% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_topography.R
\name{fgeo_topography}
\alias{fgeo_topography}
\alias{fgeo_topography.data.frame}
\alias{fgeo_topography.list}
\title{Structure topography data with mean elevation, convexity, and slope.}
\usage{
fgeo_topography(elevation, ...)

\method{fgeo_topography}{data.frame}(elevation, gridsize, xdim = NULL,
  ydim = NULL, edgecorrect = TRUE, ...)

\method{fgeo_topography}{list}(elevation, gridsize, edgecorrect = TRUE,
  ...)
}
\arguments{
\item{elevation}{One of these:
* A list with at least three elements: `col` containing
elevation data; and `xdim` and `ydim` giving plot dimensions; OR
* A dataframe containing elevation data, in which
case the parameters `xdim` and `ydim` must be provided. It may be the
element `col` of a ForestGEO elevation-list or an object of class
fgeo_elevation (see [fgeo.tool::fgeo_elevation()]).}

\item{...}{Other arguments passed to methods.}

\item{gridsize}{Number giving the size of each quadrat for which a habitat
is calculated. Commonly, `gridsize = 20`.}

\item{xdim, ydim}{(If `elevation` is a dataframe) `x` and `y` dimensions of
the plot.}

\item{edgecorrect}{Correct convexity in edge quadrats?}
}
\value{
A dataframe of subclass fgeo_topography.
}
\description{
Structure topography data with mean elevation, convexity, and slope.
}
\section{Source code}{

The main code of this function comes from [allquadratslopes()]
([source](http://bit.ly/2zikb3V)).
}

\section{Acknowledgment}{

Thanks to Jian Zhang for reporting a bug (issue 59).
}

\section{Input}{


The main input can be either the elevation list that ForestGEO delivers, or
the element \code{col} of such list -- which is a dataframe containing the
elevation data. Notice that the required arguments vary according to the main
input (the elevation list or the elevation dataframe). Whatever the input,
the dataframe containing the elevation data must have columns \code{gx} and \code{gy}
or \code{x} and \code{y}.

}

\examples{
elev_list <- fgeo.x::elevation
fgeo_topography(elev_list, gridsize = 20)

elev_df <- elev_list$col
fgeo_topography(elev_df, gridsize = 20, xdim = 320, ydim = 500)
}
\seealso{
[fgeo_habitat()].

Other functions to construct fgeo classes: \code{\link{fgeo_habitat}}

Other habitat functions: \code{\link{fgeo_habitat}},
  \code{\link{summary.tt_lst}}, \code{\link{to_df.tt_lst}},
  \code{\link{tt_test}}
}
\author{
Rick Condit.
}
\concept{functions to construct fgeo classes}
\concept{habitat functions}
