% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_tibble}
\title{Coerce objects different fgeo classes "data.frame" or "tbl" (tibble).}
\arguments{
\item{x}{An object of class tt_lst.}

\item{...}{Arguments passed on to \code{\link[base:as.data.frame]{base::as.data.frame()}} (not used in
\code{\link[=as_tibble]{as_tibble()}}).}
}
\value{
A \code{\link[base:data.frame]{base::data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\seealso{
\code{\link[base:data.frame]{base::data.frame()}}, \code{\link[tibble:as_tibble]{tibble::as_tibble()}}.
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

