% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_df.R
\name{to_df.demography_ctfs}
\alias{to_df.demography_ctfs}
\title{Dataframe objects of class "demography_impl".}
\usage{
\method{to_df}{demography_ctfs}(.x, ...)
}
\arguments{
\item{.x}{An object of class demography_impl.}

\item{...}{Other arguments passed to `to_df()`.}
}
\value{
A (tibble) dataframe.
}
\description{
Dataframe objects of class "demography_impl".
}
\examples{
\dontrun{
if (!requireNamespace("fgeo.ctfs")) {
library(fgeo.ctfs)

census1 <- fgeo.x::tree5
census2 <- fgeo.x::tree6

by <- interaction(census1$sp, census1$quadrat, sep = "__")
.x <- recruitment_impl(census1, census2, split1 = by)
head(to_df(.x))
}
}
}
\seealso{
[to_df()].
}
\concept{methods for fgeo generics}
