% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_test.R
\name{tt_test}
\alias{tt_test}
\title{Torus Translation Test to determine habitat associations of tree species.}
\usage{
tt_test(census, sp, habitat, plotdim = NULL, gridsize = NULL)
}
\arguments{
\item{census}{A dataframe; a ForestGEO _tree_ table (see details).}

\item{sp}{Character sting giving any number of species-names.}

\item{habitat}{Object giving the habitat designation for each
plot partition defined by `gridsize`. See [`fgeo_habitat()`].}

\item{plotdim, gridsize}{Plot dimensions and gridsize. If `NULL` (default)
they will be guessed, and a message will inform you of the chosen values.
If the guess is wrong, you should provide the correct values manually (and
check that your habitat data is correct).}
}
\value{
A list of matrices. You can summarize the output with [summary()] and
  convert it to a dataframe with [to_df()]. You can also view the
  result with `View(your-result)`, and reduce it from a list of matrices to a
  single matix with `Reduce(rbind, your-result)`. See examples.
}
\description{
Determine habitat-species associations with code developed by Sabrina Russo,
Daniel Zuleta, Matteo Detto, and Kyle Harms.
}
\details{
You should only try to determine the habitat association for sufficiently
abundant species. In a 50-ha plot, a minimum abundance of 50 trees/species
has been used. Also, you should use data of individual trees (i.e. a tree
table, and not a stem table with potentially multiple stems per tree. This
test only makes sense at the population level. We are interested in knowing
whether or not individuals of a species are aggregated on a habitat. Multiple
stems of an individual do not represent population level processes but
individual level processes.
}
\section{Acknowledgments}{

Nestor Engone Obiang, David Kenfack, Jennifer Baltzer, and Rutuja
Chitra-Tarak provided feedback. Daniel Zuleta provided guidance.
}

\section{Interpretation of Output}{

* `N.Hab.1`: Count of stems of the focal species in habitat 1.
* `Gr.Hab.1`: Count of instances the observed relative density of the focal
species on habitat 1 was greater than the relative density based on the TT
habitat map.
* `Ls.Hab.1`: Count of instances the observed relative density of the focal
species on habitat 1 was less than the relative density based on the TT
habitat map.
* `Eq.Hab.1`: Count of instances the observed relative density of the focal
species on habitat 1 was equal to the relative density based on the TT
habitat map.
The sum of the `Gr.Hab.x`, `Ls.Hab.x`, and `Eq.Hab.x` columns for one habitat
equals the number of 20 x20 quads in the plot.
The `Rep.Agg.Neut` columns for each habitat indicate whether the species is
significantly repelled (-1), aggregated (1), or neutrally distributed (0) on
the habitat in question.

The probabilities associated with the test for whether these patterns are
statistically significant are in the `Obs.Quantile` columns for each habitat.
Note that to calculate the probability for repelled, it is the value given,
but to calculate the probability for aggregated, it is 1 - the value given.

Values of the `Obs.Quantile` < 0.025 means that the species is repelled from
that habitat, while values of the `Obs.Quantile` > 0.975 means that the
species is aggregated on that habitat.
}

\section{References}{

Zuleta, D., Russo, S.E., Barona, A. et al. Plant Soil (2018).
\url{https://doi.org/10.1007/s11104-018-3878-0}.
}

\examples{
library(dplyr)

# Example data
tree <- luquillo_top3_sp
elevation <- fgeo.x::elevation

# Pick alive trees, of 10 mm or more
census <- filter(tree, status == "A", dbh >= 10)

# Pick sufficiently abundant species
pick <- filter(add_count(census, sp), n > 50)
species <- unique(pick$sp)

# Use your habitat data or create it from elevation data
habitat <- fgeo_habitat(elevation, gridsize = 20, n = 4)

# A list or matrices
tt_lst <- tt_test(census, species, habitat)
tt_lst

# A simple summary to help you interpret the results
summary(tt_lst)

# A combined matrix
Reduce(rbind, tt_lst)

# A dataframe
to_df(tt_lst)
}
\seealso{
[summary.tt_lst()], [to_df()], [fgeo_habitat()].

Other habitat functions: \code{\link{fgeo_habitat}},
  \code{\link{summary.tt_lst}}
}
\author{
Sabrina Russo, Daniel Zuleta, Matteo Detto, and Kyle Harms.
}
\concept{habitat functions}
